pets/requirements.txt
asgiref==3.4.1; python_version >= "3.6"
django==3.2.10; python_version >= "3.6"
djangorestframework==3.12.4; python_version >= "3.5"
pytz==2021.3; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
psycopg2==2.8.6
